+++ src/server/util/slack.js
@@ -8,23 +8,10 @@ const urljoin = require('url-join');
 /* eslint-disable no-use-before-define */
 
 module.exports = function(crowi) {
-  const { IncomingWebhook } = require('@slack/webhook');
   const { WebClient } = require('@slack/web-api');
+
   const { configManager } = crowi;
-
   const slack = {};
-
-  const postWithIwh = async(messageObj) => {
-    const webhook = new IncomingWebhook(configManager.getConfig('notification', 'slack:incomingWebhookUrl'));
-    try {
-      await webhook.send(messageObj);
-    }
-    catch (error) {
-      debug('Post error', error);
-      debug('Sent data to slack is:', messageObj);
-      throw error;
-    }
-  };
 
   const postWithWebApi = async(messageObj) => {
     const client = new WebClient(configManager.getConfig('notification', 'slack:token'));
@@ -101,7 +88,7 @@ module.exports = function(crowi) {
     return body;
   };
 
-  const prepareSlackMessageForPage = function(page, user, channel, updateType, previousRevision) {
+  slack.prepareSlackMessageForPage = (page, user, channel, updateType, previousRevision) => {
     const appTitle = crowi.appService.getAppTitle();
     const url = crowi.appService.getSiteUrl();
     let body = page.revision.body;
@@ -137,7 +124,7 @@ module.exports = function(crowi) {
     return message;
   };
 
-  const prepareSlackMessageForComment = function(comment, user, channel, path) {
+  slack.prepareSlackMessageForComment = (comment, user, channel, path) => {
     const appTitle = crowi.appService.getAppTitle();
     const url = crowi.appService.getSiteUrl();
     const body = prepareAttachmentTextForComment(comment);
@@ -171,7 +158,7 @@ module.exports = function(crowi) {
    * @param {string} attachmentBody
    * @param {string} slackChannel
   */
-  const prepareSlackMessageForGlobalNotification = async(messageBody, attachmentBody, slackChannel) => {
+  slack.prepareSlackMessageForGlobalNotification = async(messageBody, attachmentBody, slackChannel) => {
     const appTitle = crowi.appService.getAppTitle();
 
     const attachment = {
@@ -213,48 +200,25 @@ module.exports = function(crowi) {
     return text;
   };
 
-  // slack.post = function (channel, message, opts) {
   slack.postPage = (page, user, channel, updateType, previousRevision) => {
-    const messageObj = prepareSlackMessageForPage(page, user, channel, updateType, previousRevision);
+    const messageObj = slack.prepareSlackMessageForPage(page, user, channel, updateType, previousRevision);
 
     return slackPost(messageObj);
   };
 
   slack.postComment = (comment, user, channel, path) => {
-    const messageObj = prepareSlackMessageForComment(comment, user, channel, path);
+    const messageObj = slack.prepareSlackMessageForComment(comment, user, channel, path);
 
     return slackPost(messageObj);
   };
 
   slack.sendGlobalNotification = async(messageBody, attachmentBody, slackChannel) => {
-    const messageObj = await prepareSlackMessageForGlobalNotification(messageBody, attachmentBody, slackChannel);
-
+    const messageObj = await slack.prepareSlackMessageForGlobalNotification(messageBody, attachmentBody, slackChannel);
     return slackPost(messageObj);
   };
 
   const slackPost = (messageObj) => {
-    // when incoming Webhooks is prioritized
-    if (configManager.getConfig('notification', 'slack:isIncomingWebhookPrioritized')) {
-      if (configManager.getConfig('notification', 'slack:incomingWebhookUrl')) {
-        debug('posting message with IncomingWebhook');
-        return postWithIwh(messageObj);
-      }
-      if (configManager.getConfig('notification', 'slack:token')) {
-        debug('posting message with Web API');
-        return postWithWebApi(messageObj);
-      }
-    }
-    // else
-    else {
-      if (configManager.getConfig('notification', 'slack:token')) {
-        debug('posting message with Web API');
-        return postWithWebApi(messageObj);
-      }
-      if (configManager.getConfig('notification', 'slack:incomingWebhookUrl')) {
-        debug('posting message with IncomingWebhook');
-        return postWithIwh(messageObj);
-      }
-    }
+    return postWithWebApi(messageObj);
   };
 
   return slack;