+++ CHANGES.md
@@ -5,10 +5,14 @@
 * Improvement: Upgrade mongodb driver to fix [NODE-2784](https://jira.mongodb.org/browse/NODE-2784)
 * Support: Upgrade libs
     * connect-mongo
+    * i18next
     * migrate-mongo
     * mongoose
     * stream-to-promise
+    * validator
     * ws
+    * nodemailer
+    * i18next-express-middleware
 
 ## v4.2.20
 
@@ -36,14 +40,14 @@
 * Fix: Global notification to Slack does not encode spaces of page path
 * Support: Upgrade libs
     * @google-cloud/storage
-    * @slack/web-api
-    * @slack/webhook
 
 ## v4.2.17
 
 * Improvement: Invoke garbage collection when reindex all pages by elasticsearch
 * Improvement: Hide Sidebar at shared pages
 * Fix: No unsaved alert is displayed without difference the latest markdown and editor value
+* Support: Update libs
+    * eslint-config-weseek
 
 ## v4.2.16
 
@@ -59,7 +63,6 @@
     * Introduced by v4.2.8
 * Support: Upgrade libs
     * reactstrap
-
 
 
 ## v4.2.14